+++ packages/common/src/pages/index.ts
@@ -1,1 +1,2 @@
 export * from "./HubPages";
+export * from "./HubPage";