+++ packages/common/test/pages/HubPages.test.ts
@@ -1,16 +1,28 @@
+import * as portalModule from "@esri/arcgis-rest-portal";
 import { IItem } from "@esri/arcgis-rest-portal";
-
+import { MOCK_AUTH } from "../mocks/mock-auth";
+import * as modelUtils from "../../src/models";
+import * as slugUtils from "../../src/items/slugs";
+import { IRequestOptions } from "@esri/arcgis-rest-request";
 import {
   cloneObject,
   enrichPageSearchResult,
   IHubRequestOptions,
-  IHubSearchResult,
+  IModel,
 } from "../../src";
+import {
+  createPage,
+  fetchPage,
+  deletePage,
+  updatePage,
+} from "../../src/pages/HubPages";
+import { IHubPage } from "../../src/core/types/IHubPage";
 import * as FetchEnrichments from "../../src/items/_enrichments";
 
+const GUID = "f995804e9e0e42cc84187258de0b710d";
 const PAGE_ITEM: IItem = {
-  id: "f995804e9e0e42cc84187258de0b710d",
-  owner: "collabadmin7",
+  id: GUID,
+  owner: "vader",
   created: 1592615052000,
   modified: 1592615052000,
   guid: null,
@@ -65,7 +77,231 @@ const PAGE_ITEM: IItem = {
   contentOrigin: "other",
 };
 
+const PAGE_DATA = {};
+
+const PAGE_MODEL = {
+  item: PAGE_ITEM,
+  data: PAGE_DATA,
+} as IModel;
+
 describe("HubPages Module", () => {
+  describe("createPage:", () => {
+    it("works with very limited structure", async () => {
+      const slugSpy = spyOn(slugUtils, "getUniqueSlug").and.returnValue(
+        Promise.resolve("dcdev|hello-world")
+      );
+      const createSpy = spyOn(modelUtils, "createModel").and.callFake(
+        (m: IModel) => {
+          const newModel = cloneObject(m);
+          newModel.item.id = GUID;
+          return Promise.resolve(newModel);
+        }
+      );
+      const chk = await createPage(
+        { name: "Hello World", orgUrlKey: "dcdev" },
+        { authentication: MOCK_AUTH }
+      );
+
+      expect(chk.id).toBe(GUID);
+      expect(chk.name).toBe("Hello World");
+      // should ensure unique slug
+      expect(slugSpy.calls.count()).toBe(1);
+      expect(slugSpy.calls.argsFor(0)[0]).toEqual(
+        { slug: "dcdev|hello-world" },
+        "should recieve slug"
+      );
+      // should create the item
+      expect(createSpy.calls.count()).toBe(1);
+      const modelToCreate = createSpy.calls.argsFor(0)[0];
+      expect(modelToCreate.item.title).toBe("Hello World");
+      expect(modelToCreate.item.type).toBe("Hub Page");
+      expect(modelToCreate.item.properties.slug).toBe("dcdev|hello-world");
+      expect(modelToCreate.item.properties.orgUrlKey).toBe("dcdev");
+    });
+    it("works with more complete object", async () => {
+      // Note: this covers a branch when a slug is passed in
+      const slugSpy = spyOn(slugUtils, "getUniqueSlug").and.returnValue(
+        Promise.resolve("dcdev|hello-world")
+      );
+      const createSpy = spyOn(modelUtils, "createModel").and.callFake(
+        (m: IModel) => {
+          const newModel = cloneObject(m);
+          newModel.item.id = GUID;
+          return Promise.resolve(newModel);
+        }
+      );
+      const chk = await createPage(
+        {
+          name: "Hello World",
+          slug: "dcdev|hello-world", // important for coverage
+          description: "my desc",
+          orgUrlKey: "dcdev",
+        },
+        { authentication: MOCK_AUTH }
+      );
+      expect(chk.id).toBe(GUID);
+      expect(chk.name).toBe("Hello World");
+      expect(chk.description).toBe("my desc");
+      // should ensure unique slug
+      expect(slugSpy.calls.count()).toBe(1);
+      expect(slugSpy.calls.argsFor(0)[0]).toEqual(
+        { slug: "dcdev|hello-world" },
+        "should recieve slug"
+      );
+      // should create the item
+      expect(createSpy.calls.count()).toBe(1);
+      const modelToCreate = createSpy.calls.argsFor(0)[0];
+      expect(modelToCreate.item.properties.slug).toBe("dcdev|hello-world");
+      expect(modelToCreate.item.properties.orgUrlKey).toBe("dcdev");
+    });
+  });
+
+  describe("updatePage: ", () => {
+    it("updates backing model", async () => {
+      const slugSpy = spyOn(slugUtils, "getUniqueSlug").and.returnValue(
+        Promise.resolve("dcdev-wat-blarg-1")
+      );
+      const getModelSpy = spyOn(modelUtils, "getModel").and.returnValue(
+        Promise.resolve(PAGE_MODEL)
+      );
+      const updateModelSpy = spyOn(modelUtils, "updateModel").and.callFake(
+        (m: IModel) => {
+          return Promise.resolve(m);
+        }
+      );
+      const page: IHubPage = {
+        itemControl: "edit",
+        id: GUID,
+        name: "Hello World",
+        tags: ["Transportation"],
+        description: "Some longer description",
+        slug: "dcdev-wat-blarg",
+        orgUrlKey: "dcdev",
+        owner: "dcdev_dude",
+        type: "Hub Page",
+        createdDate: new Date(1595878748000),
+        createdDateSource: "item.created",
+        updatedDate: new Date(1595878750000),
+        updatedDateSource: "item.modified",
+        thumbnailUrl: "",
+        permissions: [],
+        schemaVersion: 1,
+        canEdit: false,
+        canDelete: false,
+        typeKeywords: [],
+      };
+      const chk = await updatePage(page, { authentication: MOCK_AUTH });
+      expect(chk.id).toBe(GUID);
+      expect(chk.name).toBe("Hello World");
+      expect(chk.description).toBe("Some longer description");
+      // should ensure unique slug
+      expect(slugSpy.calls.count()).toBe(1);
+      expect(slugSpy.calls.argsFor(0)[0]).toEqual(
+        { slug: "dcdev-wat-blarg", existingId: GUID },
+        "should recieve slug"
+      );
+      expect(getModelSpy.calls.count()).toBe(1);
+      expect(updateModelSpy.calls.count()).toBe(1);
+      const modelToUpdate = updateModelSpy.calls.argsFor(0)[0];
+      expect(modelToUpdate.item.description).toBe(page.description);
+      expect(modelToUpdate.item.properties.slug).toBe("dcdev-wat-blarg-1");
+    });
+  });
+
+  describe("deletePage:", () => {
+    it("deletes the item", async () => {
+      const removeSpy = spyOn(portalModule, "removeItem").and.returnValue(
+        Promise.resolve({ success: true })
+      );
+
+      const result = await deletePage("3ef", {
+        authentication: MOCK_AUTH,
+      });
+      expect(result).toBeUndefined();
+      expect(removeSpy.calls.count()).toBe(1);
+      expect(removeSpy.calls.argsFor(0)[0].authentication).toBe(MOCK_AUTH);
+      expect(removeSpy.calls.argsFor(0)[0].id).toBe("3ef");
+    });
+  });
+
+  describe("fetchPage:", () => {
+    it("gets by id, if passed a guid", async () => {
+      const getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
+        Promise.resolve(PAGE_ITEM)
+      );
+      const getItemDataSpy = spyOn(portalModule, "getItemData").and.returnValue(
+        Promise.resolve(PAGE_DATA)
+      );
+
+      const chk = await fetchPage(GUID, {
+        authentication: MOCK_AUTH,
+      });
+      expect(chk.id).toBe(GUID);
+      expect(chk.owner).toBe("vader");
+      expect(getItemSpy.calls.count()).toBe(1);
+      expect(getItemSpy.calls.argsFor(0)[0]).toBe(GUID);
+      expect(getItemDataSpy.calls.count()).toBe(1);
+      expect(getItemDataSpy.calls.argsFor(0)[0]).toBe(GUID);
+    });
+
+    it("gets without auth", async () => {
+      const getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
+        Promise.resolve(PAGE_ITEM)
+      );
+      const getItemDataSpy = spyOn(portalModule, "getItemData").and.returnValue(
+        Promise.resolve(PAGE_DATA)
+      );
+      const ro: IRequestOptions = {
+        portal: "https://gis.myserver.com/portal/sharing/rest",
+      };
+      const chk = await fetchPage(GUID, ro);
+      expect(chk.id).toBe(GUID);
+      expect(chk.owner).toBe("vader");
+      expect(chk.thumbnailUrl).toBe(
+        "https://gis.myserver.com/portal/sharing/rest/content/items/f995804e9e0e42cc84187258de0b710d/info/thumbnail/foo.png"
+      );
+      expect(getItemSpy.calls.count()).toBe(1);
+      expect(getItemSpy.calls.argsFor(0)[0]).toBe(GUID);
+      expect(getItemDataSpy.calls.count()).toBe(1);
+      expect(getItemDataSpy.calls.argsFor(0)[0]).toBe(GUID);
+    });
+
+    it("gets by slug if not passed guid", async () => {
+      const getItemBySlugSpy = spyOn(
+        slugUtils,
+        "getItemBySlug"
+      ).and.returnValue(Promise.resolve(PAGE_ITEM));
+      const getItemDataSpy = spyOn(portalModule, "getItemData").and.returnValue(
+        Promise.resolve(PAGE_DATA)
+      );
+
+      const chk = await fetchPage("dcdev-34th-street", {
+        authentication: MOCK_AUTH,
+      });
+      expect(getItemBySlugSpy.calls.count()).toBe(1);
+      expect(getItemBySlugSpy.calls.argsFor(0)[0]).toBe("dcdev-34th-street");
+      expect(getItemDataSpy.calls.count()).toBe(1);
+      expect(getItemDataSpy.calls.argsFor(0)[0]).toBe(GUID);
+      expect(chk.id).toBe(GUID);
+      expect(chk.owner).toBe("vader");
+    });
+
+    it("returns null if no id found", async () => {
+      const getItemBySlugSpy = spyOn(
+        slugUtils,
+        "getItemBySlug"
+      ).and.returnValue(Promise.resolve(null));
+
+      const chk = await fetchPage("dcdev-34th-street", {
+        authentication: MOCK_AUTH,
+      });
+      expect(getItemBySlugSpy.calls.count()).toBe(1);
+      expect(getItemBySlugSpy.calls.argsFor(0)[0]).toBe("dcdev-34th-street");
+      // This next stuff is O_o but req'd by typescript
+      expect(chk).toEqual(null as unknown as IHubPage);
+    });
+  });
+
   describe("enrichments:", () => {
     let enrichmentSpy: jasmine.Spy;
     let hubRo: IHubRequestOptions;